+++ packages/std/src/object/fromPath.ts
@@ -1,4 +1,4 @@
-import isNumber from '../types/isNumber';
+import parseFiniteNumberString from '../number/parseFiniteNumberString';
 
 /**
  * Converts the given array values into an object property path string.
@@ -8,7 +8,13 @@ import isNumber from '../types/isNumber';
  */
 export default function fromPath(arr: (string | number)[] = []): string {
   const out = arr.reduce((acc, cur) => {
-    return isNumber(cur, true) ? `${acc}[${cur}]` : `${acc}.${cur}`;
+    if (typeof cur === 'number') {
+      return Number.isFinite(cur) ? `${acc}[${cur}]` : `${acc}.${cur}`;
+    }
+
+    return !Number.isNaN(parseFiniteNumberString(cur))
+      ? `${acc}[${cur}]`
+      : `${acc}.${cur}`;
   }, '');
 
   return (out as string).slice(1);